@@ -25,11 +25,23 @@
     "@ngrx/effects": "^6.1.0",
     "@ngrx/store": "^6.1.0",
     "@ngrx/store-devtools": "^6.1.0",
+    "angularfire2": "^5.0.2",
+    "apollo-angular": "^1.3.0",
+    "apollo-angular-link-http": "^1.2.0",
+    "apollo-cache-inmemory": "^1.2.0",
+    "apollo-client": "^2.4.0",
+    "apollo-link": "^1.2.0",
+    "apollo-link-context": "^1.0.9",
+    "apollo-link-ws": "^1.0.9",
     "core-js": "^2.5.4",
+    "firebase": "^5.5.2",
+    "graphql": "^0.13.2",
+    "graphql-tag": "^2.9.2",
     "hammerjs": "^2.0.8",
     "material-design-icons": "^3.0.1",
     "ngrx-store-freeze": "^0.2.4",
     "rxjs": "^6.0.0",
+    "subscriptions-transport-ws": "^0.9.15",
     "visibilityjs": "^2.0.2",
     "zone.js": "~0.8.26"
   },
@@ -52,6 +64,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "^2.9.2"
   }
 }
