@@ -1,26 +1,44 @@
 import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
-import { ADD_ONE, AddCities, ShowCities } from '../actions';
-import { switchMap, map, catchError, tap, debounceTime } from 'rxjs/operators';
+import { ADD_ONE, AddCities, ShowCities, LOGGIN_IN_START, LoginInStart, LoginInSuccess, FETCH_CITY_START, fetchCitySuccess, REMOVE_ONE, RemoveCities } from '../actions';
+import { switchMap, map, catchError, tap, debounceTime, mergeMap } from 'rxjs/operators';
 import { of, fromEvent, interval } from 'rxjs';
 import { WeatherService } from '../../services/weather/weather.service';
+import { AuthService } from '../../services/auth/auth.service'; 
 
 @Injectable()
 export class WeatherEffects {
-  constructor(private actions$: Actions, private weatherService: WeatherService) {
+  constructor(private actions$: Actions, private weatherService: WeatherService, private authService: AuthService) {
   }
-@Effect() test$ = this.actions$.pipe(
-  ofType(ADD_ONE),
-  switchMap((action: AddCities) => 
-    this.weatherService
-      .getTest(action.payload)
-      .pipe(
-        map(c => new ShowCities(c)),
-        catchError(err => of({ type: 'ERROR', payload: err }))
+
+  @Effect() city$ = this.actions$.pipe(
+    ofType(FETCH_CITY_START),
+    mergeMap(() => this.weatherService.getCity().pipe(
+      map(({city}) => new fetchCitySuccess(city))
+    ))
+  )
+  // @Effect({dispatch: false}) AddCity$ = this.actions$.pipe(
+  //   ofType(ADD_ONE),
+  //   switchMap(({payload}: AddCities) => this.weatherService.postCity(payload)),
+  //   tap(data => console.log(data))
+  // )
+  @Effect({dispatch: false}) deleteCity$ = this.actions$.pipe(
+    ofType(REMOVE_ONE),
+    switchMap(({payload}: RemoveCities) => this.weatherService.deleteCity(payload))
+  )
+
+@Effect() auth$ = this.actions$.pipe(
+  ofType(LOGGIN_IN_START),
+  switchMap(({payload}: LoginInStart) => 
+    this.authService
+      .registerUser(payload).pipe(
+        map(res => new LoginInSuccess(res)),
+        tap(r => console.log(r)),
+        catchError((err) => of({ type: 'LOGIN_FAILED' })) 
       )
-      
   )
 )
+
 @Effect()
 resize$ = fromEvent(window, 'resize').pipe(
   debounceTime(300),
