@@ -1,17 +1,19 @@
 import {Component, OnInit, ChangeDetectionStrategy} from '@angular/core';
 import {UiService} from './services/ui/ui.service';
-
+import { Store } from '@ngrx/store';
+import * as fromRoot from './redux/reducers';
+import { fetchCity } from './redux/actions';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  
 })
 export class AppComponent implements OnInit {
   showMenu = false;
   darkModeActive: boolean;
 
-  constructor(public ui: UiService) {
+  constructor(public ui: UiService, private store: Store<fromRoot.State>) {
 
   }
 
@@ -19,6 +21,7 @@ export class AppComponent implements OnInit {
     this.ui.darkModeState.subscribe((value) => {
       this.darkModeActive = value;
     });
+    this.store.dispatch(new fetchCity());
   }
 
   toggleMenu() {
